+++ crates/objc2/src/mutability.rs
@@ -71,6 +71,12 @@ impl<IS: ?Sized> Mutability for MutableWithImmutableSuperclass<IS> {}
 impl private_mutability::Sealed for InteriorMutable {}
 impl Mutability for InteriorMutable {}
 
+impl<S: ?Sized> private_mutability::Sealed for InteriorMutableWithSubclass<S> {}
+impl<S: ?Sized> Mutability for InteriorMutableWithSubclass<S> {}
+
+impl<S: ?Sized> private_mutability::Sealed for InteriorMutableWithSuperclass<S> {}
+impl<S: ?Sized> Mutability for InteriorMutableWithSuperclass<S> {}
+
 impl private_mutability::Sealed for MainThreadOnly {}
 impl Mutability for MainThreadOnly {}
 
@@ -154,23 +160,6 @@ pub struct Mutable {
 /// - [`IsAllowedMutable`].
 /// - You are allowed to hand out pointers / references to an instance's
 ///   internal data, since you know such data will never be mutated.
-///
-///
-/// # Example
-///
-/// ```ignore
-/// unsafe impl ClassType for NSString {
-///     type Super = NSObject;
-///     type Mutability = ImmutableWithMutableSubclass<NSMutableString>;
-///     // ...
-/// }
-///
-/// unsafe impl ClassType for NSMutableString {
-///     type Super = NSString;
-///     type Mutability = MutableWithImmutableSubclass<NSString>;
-///     // ...
-/// }
-/// ```
 #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
 pub struct ImmutableWithMutableSubclass<MS: ?Sized> {
     inner: Never,
@@ -190,23 +179,6 @@ pub struct ImmutableWithMutableSubclass<MS: ?Sized> {
 /// - You are allowed to hand out pointers / references to an instance's
 ///   internal data, since you know such data will never be mutated without
 ///   the borrowchecker catching it.
-///
-///
-/// # Example
-///
-/// ```ignore
-/// unsafe impl ClassType for NSData {
-///     type Super = NSObject;
-///     type Mutability = ImmutableWithMutableSubclass<NSMutableData>;
-///     // ...
-/// }
-///
-/// unsafe impl ClassType for NSMutableData {
-///     type Super = NSData;
-///     type Mutability = MutableWithImmutableSubclass<NSData>;
-///     // ...
-/// }
-/// ```
 #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
 pub struct MutableWithImmutableSuperclass<IS: ?Sized> {
     inner: Never,
@@ -239,6 +211,70 @@ pub struct MutableWithImmutableSuperclass<IS: ?Sized> {
 #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
 pub struct InteriorMutable {
     inner: Never,
+}
+
+/// Marker type for classes that have a mutable counterpart.
+///
+/// This is effectively the same as [`InteriorMutable`], except that the type
+/// returned by `NSMutableCopying::mutableCopy` is the mutable counterpart.
+///
+/// Functionality that is provided with this:
+/// - [`IsRetainable`] -> [`ClassType::retain`].
+/// - [`IsIdCloneable`] -> [`Retained::clone`][crate::rc::Retained#impl-Clone-for-Retained<T>].
+/// - [`IsAllocableAnyThread`] -> [`ClassType::alloc`].
+///
+///
+/// # Example
+///
+/// ```ignore
+/// unsafe impl ClassType for NSString {
+///     type Super = NSObject;
+///     type Mutability = InteriorMutableWithSubclass<NSMutableString>;
+///     // ...
+/// }
+///
+/// unsafe impl ClassType for NSMutableString {
+///     type Super = NSString;
+///     type Mutability = InteriorMutableWithSuperclass<NSString>;
+///     // ...
+/// }
+/// ```
+#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
+pub struct InteriorMutableWithSubclass<Subclass: ?Sized> {
+    inner: Never,
+    subclass: PhantomData<Subclass>,
+}
+
+/// Marker type for classes that have an immutable counterpart.
+///
+/// This is effectively the same as [`InteriorMutable`], except that the type
+/// returned by `NSCopying::copy` is the immutable counterpart.
+///
+/// Functionality that is provided with this:
+/// - [`IsRetainable`] -> [`ClassType::retain`].
+/// - [`IsIdCloneable`] -> [`Retained::clone`][crate::rc::Retained#impl-Clone-for-Retained<T>].
+/// - [`IsAllocableAnyThread`] -> [`ClassType::alloc`].
+
+///
+/// # Example
+///
+/// ```ignore
+/// unsafe impl ClassType for NSData {
+///     type Super = NSObject;
+///     type Mutability = InteriorMutableWithSubclass<NSMutableData>;
+///     // ...
+/// }
+///
+/// unsafe impl ClassType for NSMutableData {
+///     type Super = NSData;
+///     type Mutability = InteriorMutableWithSuperclass<NSData>;
+///     // ...
+/// }
+/// ```
+#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
+pub struct InteriorMutableWithSuperclass<Superclass: ?Sized> {
+    inner: Never,
+    superclass: PhantomData<Superclass>,
 }
 
 /// Marker type for classes that are only safe to use from the main thread.
@@ -307,6 +343,8 @@ impl MutabilityIsIdCloneable for Root {}
 impl MutabilityIsIdCloneable for Immutable {}
 impl<MS: ?Sized> MutabilityIsIdCloneable for ImmutableWithMutableSubclass<MS> {}
 impl MutabilityIsIdCloneable for InteriorMutable {}
+impl<S: ?Sized> MutabilityIsIdCloneable for InteriorMutableWithSubclass<S> {}
+impl<S: ?Sized> MutabilityIsIdCloneable for InteriorMutableWithSuperclass<S> {}
 impl MutabilityIsIdCloneable for MainThreadOnly {}
 
 unsafe impl<T: ?Sized + ClassType> IsIdCloneable for T where T::Mutability: MutabilityIsIdCloneable {}
@@ -340,6 +378,8 @@ pub unsafe trait IsRetainable: private_traits::Sealed + IsIdCloneable {}
 trait MutabilityIsRetainable: MutabilityIsIdCloneable {}
 impl MutabilityIsRetainable for Immutable {}
 impl MutabilityIsRetainable for InteriorMutable {}
+impl<S: ?Sized> MutabilityIsRetainable for InteriorMutableWithSubclass<S> {}
+impl<S: ?Sized> MutabilityIsRetainable for InteriorMutableWithSuperclass<S> {}
 impl MutabilityIsRetainable for MainThreadOnly {}
 
 unsafe impl<T: ?Sized + ClassType> IsRetainable for T where T::Mutability: MutabilityIsRetainable {}
@@ -369,6 +409,8 @@ impl MutabilityIsAllocableAnyThread for Mutable {}
 impl<MS: ?Sized> MutabilityIsAllocableAnyThread for ImmutableWithMutableSubclass<MS> {}
 impl<IS: ?Sized> MutabilityIsAllocableAnyThread for MutableWithImmutableSuperclass<IS> {}
 impl MutabilityIsAllocableAnyThread for InteriorMutable {}
+impl<S: ?Sized> MutabilityIsAllocableAnyThread for InteriorMutableWithSubclass<S> {}
+impl<S: ?Sized> MutabilityIsAllocableAnyThread for InteriorMutableWithSuperclass<S> {}
 
 unsafe impl<T: ?Sized + ClassType> IsAllocableAnyThread for T where
     T::Mutability: MutabilityIsAllocableAnyThread
@@ -600,6 +642,22 @@ mod private_counterpart {
         type Immutable = T;
         type Mutable = T;
     }
+    impl<T, S> MutabilityCounterpartOrSelf<T> for InteriorMutableWithSubclass<S>
+    where
+        T: ClassType<Mutability = InteriorMutableWithSubclass<S>>,
+        S: ClassType<Mutability = InteriorMutableWithSuperclass<T>>,
+    {
+        type Immutable = T;
+        type Mutable = S;
+    }
+    impl<T, S> MutabilityCounterpartOrSelf<T> for InteriorMutableWithSuperclass<S>
+    where
+        T: ClassType<Mutability = InteriorMutableWithSuperclass<S>>,
+        S: ClassType<Mutability = InteriorMutableWithSubclass<T>>,
+    {
+        type Immutable = S;
+        type Mutable = T;
+    }
     impl<T: ClassType<Mutability = MainThreadOnly>> MutabilityCounterpartOrSelf<T> for MainThreadOnly {
         type Immutable = T;
         type Mutable = T;
@@ -661,6 +719,8 @@ mod tests {
         assert_traits::<ImmutableWithMutableSubclass<()>>();
         assert_traits::<MutableWithImmutableSuperclass<()>>();
         assert_traits::<InteriorMutable>();
+        assert_traits::<InteriorMutableWithSubclass<()>>();
+        assert_traits::<InteriorMutableWithSuperclass<()>>();
         assert_traits::<MainThreadOnly>();
 
         #[allow(unused)]