+++ crates/test-assembly/crates/test_ns_string/lib.rs
@@ -6,18 +6,26 @@ use objc2_foundation::{ns_string, NSString};
 // Temporary to allow testing putting string references in statics.
 // This doesn't yet compile on other platforms, but could in the future!
 #[cfg(all(target_vendor = "apple", feature = "assembly-features"))]
-#[no_mangle]
-static EMPTY: &NSString = {
-    const INPUT: &[u8] = b"";
-    objc2_foundation::__ns_string_static!(INPUT);
-    CFSTRING.as_nsstring_const()
-};
+#[repr(transparent)]
+struct StaticString(&'static NSString);
+
+#[cfg(all(target_vendor = "apple", feature = "assembly-features"))]
+unsafe impl Sync for StaticString {}
+
 #[cfg(all(target_vendor = "apple", feature = "assembly-features"))]
 #[no_mangle]
-static XYZ: &NSString = {
+static EMPTY: StaticString = {
+    const INPUT: &[u8] = b"";
+    objc2_foundation::__ns_string_static!(INPUT);
+    StaticString(CFSTRING.as_nsstring_const())
+};
+
+#[cfg(all(target_vendor = "apple", feature = "assembly-features"))]
+#[no_mangle]
+static XYZ: StaticString = {
     const INPUT: &[u8] = b"xyz";
     objc2_foundation::__ns_string_static!(INPUT);
-    CFSTRING.as_nsstring_const()
+    StaticString(CFSTRING.as_nsstring_const())
 };
 
 #[no_mangle]