+++ crates/test-ui/ui/array_iter_invalid.rs
@@ -1,29 +1,29 @@
 use objc2::rc::Retained;
-use objc2_foundation::{NSArray, NSMutableArray, NSMutableString, NSString};
+use objc2_foundation::{NSArray, NSMutableArray};
 
 fn main() {
-    let arr: Retained<NSArray<NSString>> = NSArray::new();
+    let arr: Retained<NSArray<NSArray>> = NSArray::new();
     for s in &mut arr {
-        let s: &mut NSString = s;
+        let s: &mut NSArray = s;
     }
 
-    let arr: Retained<NSArray<NSMutableString>> = NSArray::new();
+    let arr: Retained<NSArray<NSMutableArray>> = NSArray::new();
     for s in &mut arr {
-        let s: &mut NSMutableString = s;
+        let s: &mut NSMutableArray = s;
     }
 
-    let arr: Retained<NSMutableArray<NSString>> = NSMutableArray::new();
+    let arr: Retained<NSMutableArray<NSArray>> = NSMutableArray::new();
     for s in &mut arr {
-        let s: &mut NSString = s;
+        let s: &mut NSArray = s;
     }
 
     // Should succeed, included for completeness
-    let arr: Retained<NSMutableArray<NSMutableString>> = NSMutableArray::new();
+    let arr: Retained<NSMutableArray<NSMutableArray>> = NSMutableArray::new();
     for s in &mut arr {
-        let s: &mut NSString = s;
+        let s: &mut NSArray = s;
     }
 
     // RetainedIntoIterator not available for mutable children
-    let arr: Retained<NSArray<NSMutableString>> = NSArray::new();
+    let arr: Retained<NSArray<NSMutableArray>> = NSArray::new();
     for _ in arr {}
 }