+++ crates/test-ui/ui/msg_send_id_underspecified_error3.rs
@@ -1,8 +1,8 @@
 //! Test underspecified msg_send_id! errors.
 use objc2::msg_send_id;
-use objc2_foundation::NSString;
+use objc2::runtime::NSObject;
 
 fn main() {
-    let obj: &NSString;
+    let obj: &NSObject;
     let _: Result<_, _> = unsafe { msg_send_id![obj, error: _] };
 }