+++ crates/test-ui/ui/msg_send_underspecified_error.rs
@@ -1,8 +1,8 @@
 //! Test underspecified msg_send! errors.
 use objc2::msg_send;
-use objc2_foundation::NSString;
+use objc2::runtime::NSObject;
 
 fn main() {
-    let obj: &NSString;
+    let obj: &NSObject;
     let _: Result<(), _> = unsafe { msg_send![obj, error: _] };
 }