+++ framework-crates/objc2-foundation/src/tests/mutable_set.rs
@@ -2,22 +2,22 @@
 #![cfg(feature = "NSString")]
 use alloc::vec;
 
-use crate::Foundation::{self, ns_string, NSMutableSet, NSString};
+use crate::Foundation::{ns_string, NSMutableSet, NSMutableString};
 
 #[test]
 fn test_insert() {
     let mut set = NSMutableSet::new();
     assert!(set.is_empty());
 
-    assert!(set.insert_id(NSString::from_str("one")));
-    assert!(!set.insert_id(NSString::from_str("one")));
-    assert!(set.insert_id(NSString::from_str("two")));
+    assert!(set.insert(ns_string!("one")));
+    assert!(!set.insert(ns_string!("one")));
+    assert!(set.insert(ns_string!("two")));
 }
 
 #[test]
 fn test_remove() {
-    let strs = ["one", "two", "three"].map(NSString::from_str);
-    let mut set = NSMutableSet::from_id_slice(&strs);
+    let strs = [ns_string!("one"), ns_string!("two"), ns_string!("three")];
+    let mut set = NSMutableSet::from_slice(&strs);
 
     assert!(set.remove(ns_string!("one")));
     assert!(!set.remove(ns_string!("one")));
@@ -25,8 +25,8 @@ fn test_remove() {
 
 #[test]
 fn test_clear() {
-    let strs = ["one", "two", "three"].map(NSString::from_str);
-    let mut set = NSMutableSet::from_id_slice(&strs);
+    let strs = [ns_string!("one"), ns_string!("two"), ns_string!("three")];
+    let mut set = NSMutableSet::from_slice(&strs);
     assert_eq!(set.len(), 3);
 
     set.removeAllObjects();
@@ -34,12 +34,11 @@ fn test_clear() {
 }
 
 #[test]
-#[cfg(feature = "NSString")]
 fn test_into_vec() {
     let strs = vec![
-        Foundation::NSMutableString::from_str("one"),
-        Foundation::NSMutableString::from_str("two"),
-        Foundation::NSMutableString::from_str("three"),
+        NSMutableString::from_str("one"),
+        NSMutableString::from_str("two"),
+        NSMutableString::from_str("three"),
     ];
     let set = NSMutableSet::from_vec(strs);
 
@@ -49,8 +48,7 @@ fn test_into_vec() {
     }
 
     assert_eq!(vec.len(), 3);
-    let suffix = ns_string!("zero");
-    assert!(vec.iter().all(|str| str.hasSuffix(suffix)));
+    assert!(vec.iter().all(|str| str.hasSuffix(ns_string!("zero"))));
 }
 
 #[test]
@@ -58,18 +56,18 @@ fn test_extend() {
     let mut set = NSMutableSet::new();
     assert!(set.is_empty());
 
-    set.extend(["one", "two", "three"].map(NSString::from_str));
+    set.extend([ns_string!("one"), ns_string!("two"), ns_string!("three")]);
     assert_eq!(set.len(), 3);
 }
 
 #[test]
 #[cfg(feature = "NSObject")]
 fn test_mutable_copy() {
-    use Foundation::{NSMutableCopying, NSSet};
+    use crate::{NSMutableCopying, NSSet};
 
-    let set1 = NSSet::from_id_slice(&["one", "two", "three"].map(NSString::from_str));
+    let set1 = NSSet::from_slice(&[ns_string!("one"), ns_string!("two"), ns_string!("three")]);
     let mut set2 = set1.mutableCopy();
-    set2.insert_id(NSString::from_str("four"));
+    set2.insert(ns_string!("four"));
 
     assert!(set1.is_subset(&set2));
     assert_ne!(set1.mutableCopy(), set2);